Stock-Predictor AI/requirements.txt
streamlit
pandas
yfinance
textblob
requests
python-dotenv
tensorflow
numpy
scikit-learn
